Moved examples
